attribute_info: fixed data type error
The loop would use a uint8_t as the index, but the length was a uint16_t
and could be much larger than a uint8_t. If the length was >= 256, then
we would just go on reading and eventually run out of file data.

+++ src/cafebabe/attribute_info.c
@@ -37,7 +37,7 @@ cafebabe_attribute_info_init(struct cafebabe_attribute_info *a,
 	if (!a->info)
 		goto out;
 
-	for (uint8_t i = 0; i < a->attribute_length; ++i) {
+	for (uint16_t i = 0; i < a->attribute_length; ++i) {
 		if (cafebabe_stream_read_uint8(s, &a->info[i]))
 			goto out_info;
 	}
